refactor(automations): split automationExecutor into lib/automationExec modules
Pure code-move: extract fetch/scan, action executors, walk engine, trigger
matching, and dry-run engine into convex/lib/automationExec/{types,fetch,
actions,walk,matching,dryRun}.ts. automationExecutor.ts stays as the
registration facade so all internal.automationExecutor.* / api paths (crons,
eventBus, web, in-flight scheduled jobs) are unchanged; test-only exports
re-exported. All 16 registered functions and moved bodies verified
byte-identical to HEAD.
