+++ packages/backend/convex/_generated/api.d.ts
@@ -50,6 +50,12 @@ import type * as lib_activities from "../lib/activities.js";
 import type * as lib_aggregates from "../lib/aggregates.js";
 import type * as lib_assistantShared from "../lib/assistantShared.js";
 import type * as lib_auth from "../lib/auth.js";
+import type * as lib_automationExec_actions from "../lib/automationExec/actions.js";
+import type * as lib_automationExec_dryRun from "../lib/automationExec/dryRun.js";
+import type * as lib_automationExec_fetch from "../lib/automationExec/fetch.js";
+import type * as lib_automationExec_matching from "../lib/automationExec/matching.js";
+import type * as lib_automationExec_types from "../lib/automationExec/types.js";
+import type * as lib_automationExec_walk from "../lib/automationExec/walk.js";
 import type * as lib_changeTracking from "../lib/changeTracking.js";
 import type * as lib_conditionEval from "../lib/conditionEval.js";
 import type * as lib_crud from "../lib/crud.js";
@@ -194,6 +200,12 @@ declare const fullApi: ApiFromModules<{
   "lib/aggregates": typeof lib_aggregates;
   "lib/assistantShared": typeof lib_assistantShared;
   "lib/auth": typeof lib_auth;
+  "lib/automationExec/actions": typeof lib_automationExec_actions;
+  "lib/automationExec/dryRun": typeof lib_automationExec_dryRun;
+  "lib/automationExec/fetch": typeof lib_automationExec_fetch;
+  "lib/automationExec/matching": typeof lib_automationExec_matching;
+  "lib/automationExec/types": typeof lib_automationExec_types;
+  "lib/automationExec/walk": typeof lib_automationExec_walk;
   "lib/changeTracking": typeof lib_changeTracking;
   "lib/conditionEval": typeof lib_conditionEval;
   "lib/crud": typeof lib_crud;
